Moved NBTCondition and switched NBTRule to ItemRule
