+++ src/main/java/com/sabrepotato/citnbt/resources/ItemRule.java
@@ -1,5 +1,7 @@
 package com.sabrepotato.citnbt.resources;
 
+import com.sabrepotato.citnbt.resources.conditions.ItemstackCondition;
+import com.sabrepotato.citnbt.resources.conditions.NBTCondition;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
@@ -7,26 +9,31 @@ import net.minecraft.nbt.NBTTagCompound;
 import java.util.List;
 import java.util.Objects;
 
-public class NBTRule {
+public class ItemRule {
     public final List<NBTCondition> conditions;
+    public final ItemstackCondition stackCondition;
     public final ModelResourceLocation location;
 
-    public NBTRule(List<NBTCondition> conditions, ModelResourceLocation location) {
+    public ItemRule(List<NBTCondition> conditions, ModelResourceLocation location, ItemstackCondition itemstackCondition) {
         this.conditions = conditions;
         this.location = location;
+        this.stackCondition = itemstackCondition;
     }
 
     @Override
     public boolean equals(Object obj) {
-        if (!(obj instanceof NBTRule other)) return false;
+        if (!(obj instanceof ItemRule other)) return false;
         return Objects.equals(conditions, other.conditions);
     }
 
     public boolean matches(ItemStack stack) {
+        //TODO: Add nullability to each type of condition to allow skipping unused conditions
         NBTTagCompound compound = stack.getTagCompound();
         for(NBTCondition condition: conditions) {
             if (!condition.matches(compound)) return false;
         }
+        if (!stackCondition.checkConditions(stack)) return false;
+
         return true;
     }
